Refactor: RentalInfoItem 파일로 분리
