+++ src/components/item/RentalEquipmentCard.jsx
@@ -7,22 +7,8 @@ import {
   ItemCountAvailable,
   ItemCountParens,
   RentalInfoWrapper,
-  RentalInfoTitleContainer,
-  RentalInfoTitle,
-  RentalInfo,
-  RentalUserName,
-  RentalDate,
 } from './RentalEquipmentCard.styles';
-
-const RentalInfoItem = ({ user }) => {
-  return (
-    <>
-      <RentalUserName>{user.userName}</RentalUserName>
-      <RentalDate>{user.rentalStartDate}</RentalDate>
-      <RentalDate>{user.rentalEndDate}</RentalDate>
-    </>
-  );
-};
+import RentalInfoItem from './RentalInfoItem';
 
 const RentalEquipmentCard = ({ equipmentData }) => {
   return (
@@ -37,22 +23,7 @@ const RentalEquipmentCard = ({ equipmentData }) => {
         <ItemCountParens>)</ItemCountParens>
       </TitleContainer>
       <RentalInfoWrapper>
-        {equipmentData.user.length !== 0 ? (
-          <>
-            <RentalInfoTitleContainer>
-              <RentalInfoTitle>대여자</RentalInfoTitle>
-              <RentalInfoTitle>대여일</RentalInfoTitle>
-              <RentalInfoTitle>반납일</RentalInfoTitle>
-            </RentalInfoTitleContainer>
-            {equipmentData.user.map((user) => (
-              <RentalInfo key={user.studentId}>
-                <RentalInfoItem user={user} />
-              </RentalInfo>
-            ))}
-          </>
-        ) : (
-          <RentalInfoTitle color="#9D9D9D">대여자가 없습니다.</RentalInfoTitle>
-        )}
+        <RentalInfoItem user={equipmentData.user} />
       </RentalInfoWrapper>
     </CategoryItemWrapper>
   );
